+++ Thinq/ThinqUnitTests/CoPrimesUnitTest.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using System.Collections.Generic;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System.Numerics;
 
 using ThinqCore;
 
@@ -13,16 +14,16 @@ namespace ThinqUnitTests
 		[TestMethod]
 		public void TestGcdMany()
 		{
-			int[] testNumbers1 = new int[] { 6, 12, 27 };
-			int[] testNumbers2 = new int[] { 6, 12, 27, 5 };
-			int[] testNumbers3 = new int[] { 3, 7, 11, 13, 17, 19, 23 };
-			int expectedValue1 = 3;
-			int expectedValue2 = 1;
-			int expectedValue3 = 1;
+			BigInteger[] testNumbers1 = new BigInteger[] { 6, 12, 27 };
+			BigInteger[] testNumbers2 = new BigInteger[] { 6, 12, 27, 5 };
+			BigInteger[] testNumbers3 = new BigInteger[] { 3, 7, 11, 13, 17, 19, 23 };
+			BigInteger expectedValue1 = 3;
+			BigInteger expectedValue2 = 1;
+			BigInteger expectedValue3 = 1;
 
-			int resultValue1 = Coprimes.FindGCD(testNumbers1);
-			int resultValue2 = Coprimes.FindGCD(testNumbers2);
-			int resultValue3 = Coprimes.FindGCD(testNumbers3);
+			BigInteger resultValue1 = Coprimes.FindGCD(testNumbers1);
+			BigInteger resultValue2 = Coprimes.FindGCD(testNumbers2);
+			BigInteger resultValue3 = Coprimes.FindGCD(testNumbers3);
 
 			Assert.AreEqual(expectedValue1, resultValue1);
 			Assert.AreEqual(expectedValue2, resultValue2);
@@ -32,10 +33,10 @@ namespace ThinqUnitTests
 		[TestMethod]
 		public void TestLcmTwo()
 		{
-			int[] testNumbers1 = new int[] { 251, 257 };
+			BigInteger[] testNumbers1 = new BigInteger[] { 251, 257 };
 			int expectedValue1 = 64507;
 
-			int resultValue1 = Coprimes.FindLCM(testNumbers1);
+			BigInteger resultValue1 = Coprimes.FindLCM(testNumbers1);
 
 			Assert.AreEqual(expectedValue1, resultValue1);
 		}
@@ -45,16 +46,16 @@ namespace ThinqUnitTests
 		{
 			// GCD[251, 257] 1	LCM[251, 257] 64507
 
-			int[] testNumbers1 = new int[] { 6, 12, 27 };
-			int[] testNumbers2 = new int[] { 6, 12, 27, 5 };
-			int[] testNumbers3 = new int[] { 3, 7, 11, 13, 17, 19, 23 };
-			int expectedValue1 = 108;
-			int expectedValue2 = 540;
-			int expectedValue3 = 22309287;
+			BigInteger[] testNumbers1 = new BigInteger[] { 6, 12, 27 };
+			BigInteger[] testNumbers2 = new BigInteger[] { 6, 12, 27, 5 };
+			BigInteger[] testNumbers3 = new BigInteger[] { 3, 7, 11, 13, 17, 19, 23 };
+			BigInteger expectedValue1 = 108;
+			BigInteger expectedValue2 = 540;
+			BigInteger expectedValue3 = 22309287;
 
-			int resultValue1 = Coprimes.FindLCM(testNumbers1);
-			int resultValue2 = Coprimes.FindLCM(testNumbers2);
-			int resultValue3 = Coprimes.FindLCM(testNumbers3);
+			BigInteger resultValue1 = Coprimes.FindLCM(testNumbers1);
+			BigInteger resultValue2 = Coprimes.FindLCM(testNumbers2);
+			BigInteger resultValue3 = Coprimes.FindLCM(testNumbers3);
 
 			Assert.AreEqual(expectedValue1, resultValue1);
 			Assert.AreEqual(expectedValue2, resultValue2);
@@ -64,16 +65,16 @@ namespace ThinqUnitTests
 		[TestMethod]
 		public void TestCoPrimes256()
 		{
-			int coprimeTo = 256;
-			int coprimeMin = 16777216;
-			int coprimeMax = 16777700;
-			int expectedCoprimeCount = 242;
-			int expectedFirst = 16777217;
-			int expectedLast = 16777699;
+			BigInteger coprimeTo = 256;
+			BigInteger coprimeMin = 16777216;
+			BigInteger coprimeMax = 16777700;
+			BigInteger expectedCoprimeCount = 242;
+			BigInteger expectedFirst = 16777217;
+			BigInteger expectedLast = 16777699;
 
 			Coprimes coprimes = new Coprimes(coprimeTo, coprimeMin, coprimeMax);
-			List<int> resultCoprimes = coprimes.GetCoprimes().ToList();
-			int resultCoprimeCount = resultCoprimes.Count;
+			List<BigInteger> resultCoprimes = coprimes.GetCoprimes().ToList();
+			BigInteger resultCoprimeCount = resultCoprimes.Count;
 
 			Assert.AreEqual(expectedCoprimeCount, resultCoprimeCount);
 			Assert.AreEqual(expectedFirst, resultCoprimes.First());
@@ -83,16 +84,16 @@ namespace ThinqUnitTests
 		[TestMethod]
 		public void TestCoPrimes251()
 		{
-			int coprimeTo = 251;
-			int coprimeMin = 1234567000;
-			int coprimeMax = 1234568000;
-			int expectedCoprimeCount = 996;
-			int expectedFirst = 1234567000;
-			int expectedLast = 1234567999;
+			BigInteger coprimeTo = 251;
+			BigInteger coprimeMin = 1234567000;
+			BigInteger coprimeMax = 1234568000;
+			BigInteger expectedCoprimeCount = 996;
+			BigInteger expectedFirst = 1234567000;
+			BigInteger expectedLast = 1234567999;
 
 			Coprimes coprimes = new Coprimes(coprimeTo, coprimeMin, coprimeMax);
-			List<int> resultCoprimes = coprimes.GetCoprimes().ToList();
-			int resultCoprimeCount = resultCoprimes.Count;
+			List<BigInteger> resultCoprimes = coprimes.GetCoprimes().ToList();
+			BigInteger resultCoprimeCount = resultCoprimes.Count;
 
 			Assert.AreEqual(expectedCoprimeCount, resultCoprimeCount);
 			Assert.AreEqual(expectedFirst, resultCoprimes.First());
@@ -102,14 +103,14 @@ namespace ThinqUnitTests
 		[TestMethod]
 		public void TestCoPrimesEmpty()
 		{
-			int coprimeTo = 20160;
-			int coprimeMin = 6777122;
-			int coprimeMax = 6777131;
-			int expectedCoprimeCount = 0;
+			BigInteger coprimeTo = 20160;
+			BigInteger coprimeMin = 6777122;
+			BigInteger coprimeMax = 6777131;
+			BigInteger expectedCoprimeCount = 0;
 
 			Coprimes coprimes = new Coprimes(coprimeTo, coprimeMin, coprimeMax);
-			List<int> resultCoprimes = coprimes.GetCoprimes().ToList();
-			int resultCoprimeCount = resultCoprimes.Count;
+			List<BigInteger> resultCoprimes = coprimes.GetCoprimes().ToList();
+			BigInteger resultCoprimeCount = resultCoprimes.Count;
 
 			Assert.AreEqual(expectedCoprimeCount, resultCoprimeCount);
 		}
